Create a Child Rmd from a Template's Child

Sometimes a template Rmd will also have children. This function makes a copy of the template child at the given template_path and copies it to the child_rmd_folder with the same filename. Unlike create_report(), subdirectories are not created since that is the work of the main template.

create_report_child(
  template_path,
  child_rmd_folder,
  require_all_vars = TRUE,
  edit_file = TRUE,
  ...
)

Arguments

template_path
Path to the template file to use. It can point to a template Rmd in your local machine.

child_rmd_folder
Path that the template will be copied to.

require_all_vars
Variables to replace are designated as those enclosed in curly brackets in the template Rmd file. Sometimes, the template may have variables to be called upon by the glue function iteratively outside the template itself. This argument controls whether or not all of these variables must be assigned a value or not. Default: TRUE.

edit_file
If called in an interactive session, the file will be opened in R Studio if TRUE.

...
(Optional) Additional parameters that should be added in the front matter of the Rmd. Should be entered in the format of a named vector.

See also