Make DESCRIPTION

DESCRIPTION file in the given path is read and parsed into a dataframe using read_description, printed in the DESCRIPTION file format in the console. Additionally, the Imports: are read from the NAMESPACE file using read_namespace. If the DESCRIPTION file does not have an Imports section, it is returned in the console in red italics. Otherwise, it is returned in blue italics to still flag the output since this function does not compare whether the list of Imports match. This is up to the user as is updating the Remotes: section. In a similar fashion, if either the URL: or BugReports: sections are missing, they are concatenated and returned in red italics. However, they are not returned in blue italics otherwise.

make_description(path = getwd(), remote_name = "origin")

See also