Setup Tidy Git

Setup both GitHub and tidy style settings in the usethis package.

setup_tidy_git(
  initial_commit_message = "initial commit",
  organisation = NULL,
  private = FALSE,
  protocol = usethis::git_protocol(),
  host = NULL,
  steps = c("use_git", "git_vaccinate", "use_github", "use_pipe", "use_tibble",
    "use_tidy_eval", "use_tidy_style")
)

Arguments

organisation
If supplied, the repo will be created under this organisation, instead of the login associated with the GitHub token discovered for this host. The user's role and the token's scopes must be such that you have permission to create repositories in this organisation.

private
If TRUE, creates a private repository.

protocol
One of "https" or "ssh"

host
GitHub host to target, passed to the .api_url argument of gh::gh(). If unspecified, gh defaults to "https://api.github.com", although gh's default can be customised by setting the GITHUB_API_URL environment variable.

For a hypothetical GitHub Enterprise instance, either "https://github.acme.com/api/v3" or "https://github.acme.com" is acceptable.

See also